friendships controller, tested
