+++ app/views/friendships/_request.html.erb
@@ -11,10 +11,10 @@
     </div>
     <div class="pull-right">
       <div>
-        <%= link_to 'Accept', accept_friend_request_user_path(id: current_user.id, friend_id: request.user.id), method: :post, class: 'btn btn-success' %>
+        <%= link_to 'Accept', accept_friend_request_path(user_id: current_user.id, friend_id: request.user.id), method: :post, class: 'btn btn-success' %>
       </div>
       <div>
-        <%= link_to 'Reject', reject_friend_request_user_path(id: current_user.id, friend_id: request.user.id), method: :delete, class: 'btn btn-danger mt-1' %>
+        <%= link_to 'Reject', reject_friend_request_path(user_id: current_user.id, friend_id: request.user.id), method: :delete, class: 'btn btn-danger mt-1' %>
       </div>
     </div>
   </div>
