+++ app/views/friendships/friend_requests.html.erb
