+++ app/views/layouts/_navigation.html.erb
@@ -17,11 +17,11 @@
       </ul>
       <ul class="nav navbar-nav navbar-right">
         <% if current_user.friend_requests_count.zero? %>
-          <li><%= link_to 'Friend requests', friend_requests_user_path(current_user) %></li>
+          <li><%= link_to 'Friend requests', friend_requests_path(user_id: current_user.id) %></li>
         <% else %>
-          <li><%= link_to "Friend requests <span class='badge'>#{current_user.friend_requests_count}</span>".html_safe, friend_requests_user_path(current_user) %> </li>
+          <li><%= link_to "Friend requests <span class='badge'>#{current_user.friend_requests_count}</span>".html_safe, friend_requests_path(user_id: current_user.id) %> </li>
         <% end %>
-        <li><%= link_to 'Friends', friends_user_path(current_user) %></li>
+        <li><%= link_to 'Friends', friends_path(user_id: current_user.id) %></li>
         <li class="dropdown">
           <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Account <span class="caret"></span></a>
           <ul class="dropdown-menu">
