+++ app/views/shared/index.html.erb
