+++ app/views/users/_user.html.erb
@@ -13,7 +13,7 @@
     </dl>
     <% if current_user != user && !current_user.real_friends.include?(user) && !current_user.friends.include?(user) %>
       <div class="cleared pull-right">
-        <%= link_to 'Ask friendship', ask_friendship_user_path(id: current_user.id, friend_id: user.id), method: :post, class: 'btn btn-primary' %>
+        <%= link_to 'Ask friendship', ask_friendship_path(user_id: current_user.id, friend_id: user.id), method: :post, class: 'btn btn-primary' %>
       </div>
     <% end %>
   </div>
