+++ config/routes.rb
@@ -9,20 +9,22 @@ Rails.application.routes.draw do
 
   resources :users, only: [:index, :show, :edit, :update] do
     get 'posts', on: :member, to: 'users#user_posts'
-    get 'friend_requests', on: :member, to: 'users#friend_requests'
-    get 'friends', on: :member, to: 'users#friends_index'
-    post 'accept_friend_request', on: :member, to: 'users#accept_friend_request'
-    post 'ask_friendship', on: :member, to: 'users#ask_friendship'
-    delete 'reject_friend_request', on: :member, to: 'users#reject_friend_request'
-    delete 'cancel_friendship', on: :member, to: 'users#cancel_friendship'
   end
 
   # likes
   post 'like_post', to: 'likes#like_post'
   delete 'unlike_post', to: 'likes#unlike_post'
 
-  # posts
+  # comments
   post 'comment_post', to: 'comments#comment_post'
+
+  # friendships
+  get 'friend_requests', to: 'friendships#friend_requests'
+  get 'friends', to: 'friendships#friends_index'
+  post 'accept_friend_request', to: 'friendships#accept_friend_request'
+  post 'ask_friendship', to: 'friendships#ask_friendship'
+  delete 'reject_friend_request', to: 'friendships#reject_friend_request'
+  delete 'cancel_friendship', to: 'friendships#cancel_friendship'
 
   resources :posts
 end
